requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==5.2.4
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-environ==0.12.0
django-oauth-toolkit==3.0.1
djangorestframework==3.16.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
